move tracer tests to tests/tracer
