+++ pyproject.toml
@@ -178,23 +178,8 @@ exclude = '''
       | test_utils.py
     )
     | profiling/_ast_test_file.py
-    | propagation/test_utils.py
     | subprocesstest.py
-    | test_api.py
-    | test_compat.py
-    | test_context.py
-    | test_encoders.py
-    | test_filters.py
-    | test_global_config.py
-    | test_helpers.py
-    | test_hook.py
-    | test_instance_config.py
-    | test_payload.py
-    | test_pin.py
-    | test_sampler.py
-    | test_span.py
-    | test_tracer.py
-    | test_worker.py
+    | tracer/
     | unit/
     (
       test_settings.py
