+++ tests/contrib/aiopg/test.py
@@ -14,7 +14,7 @@ from ddtrace import Pin
 # testing
 from tests.opentracer.utils import init_tracer
 from tests.contrib.config import POSTGRES_CONFIG
-from tests.test_tracer import get_dummy_tracer
+from tests.tracer.test_tracer import get_dummy_tracer
 from tests.contrib.asyncio.utils import AsyncioTestCase, mark_asyncio
 from ...subprocesstest import run_in_subprocess
 from ...utils import assert_is_measured
