+++ tests/contrib/falcon/test_distributed_tracing.py
@@ -1,5 +1,5 @@
 from falcon import testing
-from tests.test_tracer import get_dummy_tracer
+from tests.tracer.test_tracer import get_dummy_tracer
 
 from .app import get_app
