+++ tests/contrib/flask_autopatch/test_flask_autopatch.py
@@ -7,7 +7,7 @@ from ddtrace.vendor import wrapt
 from ddtrace.ext import http
 from ddtrace import Pin
 
-from ...test_tracer import get_dummy_tracer
+from tests.tracer.test_tracer import get_dummy_tracer
 from ...utils import assert_span_http_status_code, assert_is_measured
