+++ tests/contrib/flask_cache/test_wrapper_safety.py
@@ -13,7 +13,7 @@ from redis.exceptions import ConnectionError
 import pytest
 
 # testing
-from ...test_tracer import DummyWriter
+from tests.tracer.test_tracer import DummyWriter
 
 
 class FlaskCacheWrapperTest(unittest.TestCase):
