+++ tests/contrib/gevent/test_tracer.py
@@ -10,7 +10,7 @@ from ddtrace.ext.priority import USER_KEEP
 from unittest import TestCase
 from opentracing.scope_managers.gevent import GeventScopeManager
 from tests.opentracer.utils import init_tracer
-from tests.test_tracer import get_dummy_tracer
+from tests.tracer.test_tracer import get_dummy_tracer
 
 from .utils import silence_errors
