+++ tests/contrib/jinja2/test_jinja2.py
@@ -5,7 +5,7 @@ import jinja2
 
 from ddtrace import Pin, config
 from ddtrace.contrib.jinja2 import patch, unpatch
-from tests.test_tracer import get_dummy_tracer
+from tests.tracer.test_tracer import get_dummy_tracer
 from ...base import BaseTracerTestCase
 from ...utils import assert_is_measured, assert_is_not_measured
