+++ tests/contrib/mako/test_mako.py
@@ -8,7 +8,7 @@ from mako.runtime import Context
 from ddtrace import Pin
 from ddtrace.contrib.mako import patch, unpatch
 from ddtrace.compat import StringIO, to_unicode
-from tests.test_tracer import get_dummy_tracer
+from tests.tracer.test_tracer import get_dummy_tracer
 from ...base import BaseTracerTestCase
 from ...utils import assert_is_measured
