+++ tests/contrib/mongoengine/test_backwards.py
@@ -3,7 +3,7 @@ ensure old interfaces exist and won't break things.
 """
 import mongoengine
 
-from tests.test_tracer import get_dummy_tracer
+from tests.tracer.test_tracer import get_dummy_tracer
 from tests.contrib import config
