+++ tests/contrib/mysql/test_backwards_compatibility.py
@@ -1,6 +1,6 @@
 
 from ddtrace.contrib.mysql import get_traced_mysql_connection
-from tests.test_tracer import get_dummy_tracer
+from tests.tracer.test_tracer import get_dummy_tracer
 from tests.contrib import config
