+++ tests/contrib/pymongo/test.py
@@ -15,7 +15,7 @@ from ddtrace.contrib.pymongo.patch import patch, unpatch
 from tests.opentracer.utils import init_tracer
 from ..config import MONGO_CONFIG
 from ...base import BaseTracerTestCase, override_config
-from ...test_tracer import get_dummy_tracer
+from tests.tracer.test_tracer import get_dummy_tracer
 from ...utils import assert_is_measured
