+++ tests/contrib/pymysql/test_backwards_compatibility.py
@@ -1,5 +1,5 @@
 from ddtrace.contrib.pymysql import get_traced_pymysql_connection
-from tests.test_tracer import get_dummy_tracer
+from tests.tracer.test_tracer import get_dummy_tracer
 from tests.contrib import config
