+++ tests/contrib/vertica/test_vertica.py
@@ -14,7 +14,7 @@ from ddtrace.ext import errors
 from tests.base import BaseTracerTestCase
 from tests.contrib.config import VERTICA_CONFIG
 from tests.opentracer.utils import init_tracer
-from tests.test_tracer import get_dummy_tracer
+from tests.tracer.test_tracer import get_dummy_tracer
 from ...utils import assert_is_measured
 
 TEST_TABLE = 'test_table'
