+++ tests/tracer/test_context.py
@@ -3,8 +3,8 @@ import logging
 import mock
 import threading
 
-from .base import BaseTestCase
-from tests.test_tracer import get_dummy_tracer
+from tests.base import BaseTestCase
+from .test_tracer import get_dummy_tracer
 
 import pytest
