+++ tests/tracer/test_global_config.py
@@ -6,8 +6,8 @@ import pytest
 from ddtrace import config as global_config
 from ddtrace.settings import Config
 
+from ..utils import override_env
 from .test_tracer import get_dummy_tracer
-from .utils import override_env
 
 
 class GlobalConfigTestCase(TestCase):
