+++ tests/tracer/test_hook.py
