+++ tests/tracer/test_hooks.py
