+++ tests/tracer/test_payload.py
@@ -4,7 +4,7 @@ from ddtrace.encoding import Encoder, JSONEncoder
 from ddtrace.payload import Payload, PayloadFull
 from ddtrace.span import Span
 
-from .base import BaseTracerTestCase
+from tests.base import BaseTracerTestCase
 
 import pytest
