+++ tests/tracer/test_rand.py
