+++ tox.ini
@@ -461,7 +461,7 @@ passenv=TEST_*
 
 commands =
 # run only essential tests related to the tracing client
-    tracer: pytest {posargs} --ignore="tests/contrib" --ignore="tests/integration" --ignore="tests/commands" --ignore="tests/opentracer" --ignore="tests/unit" --ignore="tests/internal" --ignore="tests/profiling"
+    tracer: pytest {posargs} tests/tracer
     profile: python -m tests.profiling.run pytest --capture=no --verbose {posargs} tests/profiling
 # run only the `ddtrace.internal` tests
     internal: pytest {posargs} tests/internal
